He arreglado un par de cosas y excepciones varias

+++ slideshare.py
@@ -1,26 +1,39 @@
+# Iba a comentar este código, pero me ha dado mucha pereza así que así se queda
 try:
     import requests
 except:
-    install("requests")
+    print("Es necesario instalar el módulo requests\nLo puedes instalar con \"pip install requests\"")
+    raise SystemExit
+from genericpath import exists
 from gettext import install
 import os
 import shutil
 try:
     from PIL import Image
 except:
-    install("Pillow")
-
+    print("Es necesario instalar el módulo PIL\nLo puedes instalar con \"pip install Pillow\"")
+    raise SystemExit(0)
 import pathlib
-
-
 
 def downloadImages():
     path = pathlib.Path(__file__).parent.absolute()
-    try:
-        shutil.rmtree(os.path.join(path, "images"))
-    except:
+
+    if exists("images"):
+        dir = os.listdir("images")
+        if len(dir) != 0:
+            print()
+            texto = input("El directorio \"images\" no está vacio ¿Desea borrar su contenido?[S/N]: ")
+            if texto.capitalize() == "S":
+                shutil.rmtree(os.path.join(path, "images"))
+                os.mkdir("images")
+            else:
+                print("No se puede continuar si el directorio \"images\" no está vacio")
+                raise SystemExit(0)
+    else:
         print("Creando directorio de imágenes...")
-    os.mkdir("images")
+        os.mkdir("images")
+    
+    print("Descargarndo imágenes de SlideShare...")
     for i in range(1,22):
         url = "https://image.slidesharecdn.com/plandecontingencia-140605102313-phpapp01/95/plan-de-contingencia-" + str(i) + "-638.jpg?cb=1401963929"
 
@@ -31,6 +44,7 @@ def downloadImages():
             handler.write(img_data)
 
 def makePDF():
+    print("Creando PDF...")
     path = pathlib.Path(__file__).parent.absolute()
     #imagesJPG = list()
     imagesRGB = list()
@@ -43,13 +57,11 @@ def makePDF():
     pdfPath = os.path.join(path, "images")
     pdfPath = os.path.join(pdfPath, "image-0.jpg")
     imagesRGB[0].save("result.pdf", save_all=True, append_images=imagesRGB)
+    print("Se ha creado el PDF \"result.pdf\"")
 
 def main():
-    path = pathlib.Path(__file__).parent.absolute()
-    # print(os.path.join(path,"images"))
     downloadImages()
     makePDF()
-
 
 if __name__ == "__main__":
     main()
